requirements.txt
Django==4.0.3
django-debug-toolbar==3.2.4
django-extensions==3.1.5
django-crispy-forms==1.14.0
crispy-bootstrap5==0.6
django-registration-redux==2.10
django-imagekit==4.1.0
Pillow==9.1.0

Faker==13.3.4
ipython==8.2.0
black==22.3.0
